update on the view password
the js code for the view password works but the blocker is the function does not work throughout the code it functions well on two password input forms but does not work on one. the proffered solution is to write two different js codes to fix the problem instead of having one block code to solve the problem.